Resolved Issue #37
Button for listing all class contents moved to main page, per client request in issue #37.
