+++ src/velcro/LandingPage.java
@@ -15,7 +15,10 @@ import java.awt.Font;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import javax.swing.JTextField;
+import javax.swing.table.DefaultTableModel;
 import javax.swing.JPasswordField;
+import javax.swing.JScrollPane;
+import javax.swing.JTable;
 import javax.swing.JButton;
 import javax.swing.JLabel;
 import javax.swing.JOptionPane;
@@ -101,7 +104,7 @@ public class LandingPage {
 		// Button to navigate to Relationships page
 		btnNewButton_2 = new JButton("Relationships");
 		btnNewButton_2.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		btnNewButton_2.setBounds(76, 190, 207, 57);
+		btnNewButton_2.setBounds(76, 164, 207, 57);
 		btnNewButton_2.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				homepage.dispose();
@@ -113,7 +116,7 @@ public class LandingPage {
 		// Button to navigate to Save page
 		btnNewButton_3 = new JButton("Save");
 		btnNewButton_3.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		btnNewButton_3.setBounds(358, 190, 207, 57);
+		btnNewButton_3.setBounds(358, 164, 207, 57);
 		btnNewButton_3.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				homepage.dispose();
@@ -125,7 +128,7 @@ public class LandingPage {
 		// Button to navigate to Load page
 		btnNewButton_4 = new JButton("Load");
 		btnNewButton_4.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		btnNewButton_4.setBounds(358, 319, 207, 57);
+		btnNewButton_4.setBounds(358, 272, 207, 57);
 		btnNewButton_4.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				homepage.dispose();
@@ -134,10 +137,70 @@ public class LandingPage {
 		});
 		homepage.getContentPane().add(btnNewButton_4);
 
+		// Button to display class list and contents.
+		JButton btnClassList = new JButton("All Class Contents");
+		btnClassList.setFont(new Font("Tahoma", Font.PLAIN, 15));
+		btnClassList.setBounds(76, 275, 207, 54);
+		btnClassList.addActionListener(new ActionListener() {
+			public void actionPerformed(ActionEvent e) {
+				// Aborts if classList is empty
+				if (thisInstance.classList == null || thisInstance.classList.length == 0) {
+					JOptionPane.showMessageDialog(homepage, "Please enter classes first.");
+					return;
+				}
+
+				// Sets up table for pop-up
+				DefaultTableModel model = new DefaultTableModel();
+				model.addColumn("Name");
+				model.addColumn("Attribute");
+				model.addColumn("Relationship Source");
+				model.addColumn("Relationship Destination");
+				boolean first = true;
+				for (int h = 0; h < thisInstance.classList.length; h++) {
+					// Adds a blank line between classes.
+					if (first) {
+						first = false;
+					} else {
+						model.addRow(new Object[] { " ", " ", " ", " "});
+					}
+					Classes thisClass = thisInstance.classList[h];
+					// Adds class name
+					model.addRow(new Object[] { thisClass.getName() });
+					// Adds all class's attributes
+					if (thisClass.attributeList != null) {
+						if (thisClass.attributeList.length != 0) {
+							for (int i = 0; i < thisClass.attributeList.length; i++) {
+								model.addRow(new Object[] { " ", thisClass.attributeList[i].getName() });
+							}
+						}
+					}
+					// Adds all class's relationships
+					if (thisClass.relationshipList != null) {
+						if (thisClass.relationshipList.length != 0) {
+							for (int i = 0; i < thisClass.relationshipList.length; i++) {
+								model.addRow(new Object[] { " ", " ", thisClass.relationshipList[i].getSource(),
+										thisClass.relationshipList[i].getDestination() });
+
+							}
+						}
+					}
+
+				}
+				// Adjusting table so the headers are fully visible
+				JTable table = new JTable(model);
+				table.getColumnModel().getColumn(2).setPreferredWidth(110);
+				table.getColumnModel().getColumn(3).setPreferredWidth(135);
+				JOptionPane.showMessageDialog(null, new JScrollPane(table));
+
+			}
+		});
+		homepage.getContentPane().add(btnClassList);
+		
+		
 		// Exit button
 		btnNewButton_6 = new JButton("Exit");
 		btnNewButton_6.setFont(new Font("Tahoma", Font.PLAIN, 16));
-		btnNewButton_6.setBounds(76, 319, 207, 57);
+		btnNewButton_6.setBounds(221, 374, 207, 57);
 		btnNewButton_6.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				homepage.dispose();
